TypeDBClient.AttributeType

TypeDBClient.RemoteConcept

TypeDBClient.RemoteRule

Base.delete!

Base.push!

TypeDBClient._is_stream_respart_done

TypeDBClient.as_remote

TypeDBClient.as_remote

TypeDBClient.batch_requests

TypeDBClient.bytes

TypeDBClient.collect_result

TypeDBClient.copy_to_proto

TypeDBClient.delete

TypeDBClient.delete

TypeDBClient.delete

TypeDBClient.delete

TypeDBClient.make_pulse_request

TypeDBClient.process_direct_requests

TypeDBClient.safe_close

TypeDBClient.single_request

TypeDBClient.stream_request

TypeDBClient.AttributeType — Type

AttributeType{V} <: AbstractAttributeType

Type parameter V contains the value type of an attribute.

(OBJECT = 0, BOOLEAN = 1, LONG = 2, DOUBLE = 3, STRING = 4, DATETIME = 5)

TypeDBClient.RemoteConcept — Type

Remote

Wrapper type that encapsulates a concept and transaction.

TypeDBClient.RemoteRule — Type

Remote

Wrapper type that encapsulates a Rule and transaction.

Base.delete! — Method

delete!(resp_collector::ResponsCollector, id::Bytes)

The function will close the collecting result channel and remove this from the the resonse collector. Attention! Don't remove a result_channel manually from the Dictionary. This will not be thread safe.

Base.push! — Method

push!(resp_collector::ResponsCollector, request::T) where {T<:Proto.ProtoType}

Function is meant to give back the result_channel in which the results for one request will be collected. Attention! Don't put a new Id manually on the ResponsCollector. It wouldn't be thread safe

TypeDBClient._is_stream_respart_done — Method

function isstreamrespartdone(req_result::Proto.ProtoType) This function decides how to treat the result. It returns whether it should push the request to the answers and if it should break the retrieving loop.

TypeDBClient.as_remote — Method

as_remote(x, t)

Create a Remote(@ref) object for a concept x with a transaction t.

TypeDBClient.as_remote — Method

as_remote(x, t)

Create a Remote(@ref) object for a concept x with a transaction t.

TypeDBClient.batch_requests — Method

function batchrequests(inchannel::Channel{Proto.ProtoType}, outchannel::Channel{Proto.TransactionClient}) This function contains the whole logic for batching incomming requests. The inner runner function will be called every x ms and will send the collected request in one Transaction_Client message to teh server.

TypeDBClient.bytes — Method

bytes(x)

Convert x to Vector{UInt8}.

TypeDBClient.collect_result — Method

function collectresult(reschannel::Channel{T}) where {T<:ProtoProtoType} The function will be called for each single request. She works until the whole result set will be collected.

TypeDBClient.copy_to_proto — Method

copytoproto(fromobject, toproto_struct::Type{T}) where {T<: ProtoType} perform the conversation from a normal struct to a proto struct. The prerequisite for this is the naming of the variable of the normal struct according to the proto struct which is to be built.

TypeDBClient.delete — Function

delete(transaction::AbstractCoreTransaction, query::AbstractString, options = Proto.Options())

To delete something without using certain functions you can use a TypeQL string put this as an argument to the delete function.

TypeDBClient.delete — Method

delete(r::RemoteConcept{<:AbstractThing})

To delete an Entity, Relation or Attribute this have to be packed with the transaction to a RemoteConcept. To accomplish this see the function as_remote.

TypeDBClient.delete — Method

delete(remote::RemoteRule)

To delete a type in the database pack the type with the transaction to a RemoteRule.

TypeDBClient.delete — Method

delete(r::RemoteConcept{C,T}) where {C <: AbstractThingType,T <: AbstractCoreTransaction}

To delete a type in the database pack the type with the transaction to a RemoteConcept. Be aware that a type can only be deleted if no Entity, Attribute or Relation is in the database which is based on this type.

TypeDBClient.make_pulse_request — Method

make_pulse_request(session::T) where {T<:AbstractCoreSession}

This function make a pulse request to keep the session alive.

TypeDBClient.process_direct_requests — Method

function processdirectrequests(inchannel::Channel{Proto.ProtoType}, outchannel::Channel{Proto.Transaction_Client}) This function process the incoming request directly to the server

TypeDBClient.safe_close — Method

safe_close(source_to_close)

Close a given resource which implementes the close functionalety safely and only logs potential errors

TypeDBClient.single_request — Method

function singlerequest(bidirectstream::BidirectionalStream, request::T, batch::Bool) where {T<: Proto.ProtoType} This function process one single request and give back the results from the server to the calling functions. It is an intern function.

TypeDBClient.stream_request — Method

function streamrequest(bidirectstream::BidirectionalStream, request::T, batch::Bool) where {T<: Proto.ProtoType} Here we let the user decide what to do with the result channel. The returned value is the pure result_channel.

API reference to be filled